The interface of basic data management should be judged by the administrator. (#5108)
* Add administrator judgment

* Add jdbc params

* Add file path judgment

* add throw Exception

* code format

code format

* remove no use file

* code optimize

* use feign requestInterceptor to fixed server

* add gateway ip priority load balancer

* Fix Rpc bug

* Support automatic retry for important RPC requests

* Fix build issue

* code format

+++ linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/basedatamanager/server/restful/DatasourceEnvRestfulApi.java
@@ -97,7 +97,12 @@ public class DatasourceEnvRestfulApi {
       httpMethod = "DELETE")
   @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
   public Message remove(HttpServletRequest request, @PathVariable("id") Long id) {
-    ModuleUserUtils.getOperationUser(request, "Remove a Datasource Env Record,id:" + id.toString());
+    String username =
+        ModuleUserUtils.getOperationUser(
+            request, "Remove a Datasource Env Record,id:" + id.toString());
+    if (!Configuration.isAdmin(username)) {
+      return Message.error("User '" + username + "' is not admin user[非管理员用户]");
+    }
     boolean result = datasourceEnvService.removeById(id);
     return Message.ok("").data("result", result);
   }
